# internal/transport/httpapi/docs/openapi.json
{"openapi":"3.0.3","info":{"title":"go-whatsapp-ai-integration-POC","version":"1.0.0","description":"POC API for receiving WhatsApp messages, generating a Gemini reply, and responding in real time."},"paths":{"/healthz":{"get":{"summary":"Health check","responses":{"200":{"description":"Application status","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HealthResponse"}}}}}}},"/simulate":{"post":{"summary":"Simulate a chat message without WhatsApp","requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SimulationRequest"},"example":{"phone_number":"5511999999999","message":"What documents do I need to renew my registration?"}}}},"responses":{"200":{"description":"Generated reply","content":{"application/json":{"schema":{"$ref":"#/components/schemas/SimulationResponse"}}}},"400":{"description":"Validation error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"403":{"description":"Phone number is not allowed","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}}}}},"/webhook":{"get":{"summary":"Verify the Meta webhook","parameters":[{"name":"hub.mode","in":"query","required":true,"schema":{"type":"string","example":"subscribe"}},{"name":"hub.verify_token","in":"query","required":true,"schema":{"type":"string"}},{"name":"hub.challenge","in":"query","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"Webhook verified"},"403":{"description":"Invalid verification token","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}}}},"post":{"summary":"Receive WhatsApp webhook notifications","requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/WhatsAppWebhookNotification"}}}},"responses":{"200":{"description":"Notification accepted","content":{"application/json":{"schema":{"$ref":"#/components/schemas/WebhookResponse"}}}},"400":{"description":"Invalid payload","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}}}}},"/swagger/openapi.json":{"get":{"summary":"OpenAPI document","responses":{"200":{"description":"OpenAPI specification"}}}}},"components":{"schemas":{"HealthResponse":{"type":"object","properties":{"status":{"type":"string"},"whatsapp_sender_configured":{"type":"boolean"},"whatsapp_webhook_configured":{"type":"boolean"},"gemini_configured":{"type":"boolean"},"gemini_model":{"type":"string"}},"required":["status","whatsapp_sender_configured","whatsapp_webhook_configured","gemini_configured","gemini_model"]},"SimulationRequest":{"type":"object","properties":{"phone_number":{"type":"string"},"message":{"type":"string"}},"required":["phone_number","message"]},"SimulationResponse":{"type":"object","properties":{"phone_number":{"type":"string"},"input_message":{"type":"string"},"reply_message":{"type":"string"}},"required":["phone_number","input_message","reply_message"]},"WebhookResponse":{"type":"object","properties":{"status":{"type":"string"}},"required":["status"]},"ErrorResponse":{"type":"object","properties":{"error":{"type":"string"}},"required":["error"]},"WhatsAppWebhookNotification":{"type":"object","properties":{"entry":{"type":"array","items":{"$ref":"#/components/schemas/WhatsAppWebhookEntry"}}}},"WhatsAppWebhookEntry":{"type":"object","properties":{"changes":{"type":"array","items":{"$ref":"#/components/schemas/WhatsAppWebhookChange"}}}},"WhatsAppWebhookChange":{"type":"object","properties":{"value":{"$ref":"#/components/schemas/WhatsAppWebhookValue"}}},"WhatsAppWebhookValue":{"type":"object","properties":{"contacts":{"type":"array","items":{"$ref":"#/components/schemas/WhatsAppContact"}},"messages":{"type":"array","items":{"$ref":"#/components/schemas/WhatsAppMessage"}}}},"WhatsAppContact":{"type":"object","properties":{"wa_id":{"type":"string"}}},"WhatsAppMessage":{"type":"object","properties":{"from":{"type":"string"},"type":{"type":"string"},"text":{"$ref":"#/components/schemas/WhatsAppTextEnvelope"}}},"WhatsAppTextEnvelope":{"type":"object","properties":{"body":{"type":"string"}}}}}}